[internal/common]: Move testing/util to maps, usage outside of tests already
